+++ bika/lims/upgrade/v01_03_003.py
@@ -133,6 +133,9 @@ INDEXES_TO_ADD = [
 
     # Default listing_searchable_text index adapter for setup_catalog
     ("bika_setup_catalog", "category_uid", "KeywordIndex"),
+
+    # Allows to search ARReports where the current Sample UID is included
+    ("portal_catalog", "sample_uid", "KeywordIndex"),
 ]
 
 INDEXES_TO_REMOVE = [